requirements.txt
pandas
fastapi
langchain
tiktoken
uvicorn
python-dotenv
shortuuid
torch
chromadb
sentence-transformers
unstructured
python-docx
bson
httpx==0.24.1
redis
pymongo==4.4.0
